Add Example For Space Complexity
Add example for Space Complexity which is how much additional memory do
we need to allocate to run the code in our algorithm, technically this
is Space Complexity? However, in this course, we are going to look at
the algorithms which do not include the space taken up by the inputs,
which are called Auxiliary Space Complexity. We are going to focus on
what happens inside the algorithm, not the space the input takes up.
That said, from here on, going to be referred to as Space Complexity as
that is in the course.