Add Push Function to DoublyLinkedList Class
Add the Push function to the DoublyLinkedList class to update the tail
to point to a new node when the developer wants to add a node to the end
of the list. Also, updates the previous to point to the old tail so that
the developer can go back to the original tail.