Add Example of Stack Data Struct Using Linked List
Add an example of a Stack data structure using a linked list to show
that a linked list works well with a stack. Since a Stack is adding
elements to the start or end of a Stack as long as the adding and
removing is the same and that a Stack needs to be the inconsistent time,
it is best to use `unshift` / `shift`. As adding Nodes to be back, the
end is you will need to get to the end of the list if the Linked List is
a Singly, which this Stack is currently using, to get previous to the
end and remove a Node from there.