Add Example of Qunue Data Struct Using Linked List
Add an example of a Queue data structure using a linked list to show
that a linked list works well with a Queue. Since a Queue is adding
elements to the end of a Queue and removing the elements at the end, it
is best to use `push` / `shift`, which is called Enqueue (`push`) and
Dequeue (`shift`).