Add Binary Search Tree Class & Node Class
Add a Binary Search Tree class and its Node class to show how a Binary
Search Tree class and a Node class work together. A Binary Search Tree
can only have a node with two or fewer child nodes, and besides, the
left side of the node is smaller value, then its parent and the right
side has the amount that is larger than its parent. So the left side of
the tree should not have a more considerable value, then the right side
of the tree.