Add Tree Traversal BFS to Binary Search Tree Class
Add Breadth-first Search (BFS) to the Binary Search Tree class to show
how to do tree traversal, such as BST, as not all trees organize.
Because of this, we need an algorithm to search the Tree data structure
since they are involved. There can be many ways to do those searches BFS
is one of the main ways to search in a Tree data structure.

Also, since we already have a tree class setup, we have to use the BST
class to do tree traversal; however, this can apply to any tree data
structure type.