Add Tree Traversal DFS PreOrder to BST Class
Add Tree Traversal Depth-first Search (DFS) to the Binary Search Tree
class, a search that does an up and down fashion. In a Depth-first
traversal, there are three different ways of doing it.

The PreOrder way is that the search goes from the root, then left
checking each node on the left side, then moving up and inspecting each
parent's right-side until there is a left node, then reviewing those
nodes before continuing. To the root node, doing the same, yet this time
on the right side of the tree.