Add Tree Traversal DFS InOrder to BST Class
The InOrder way is that the search there is the left-most then goes it,
parent, then any siblings before moving to the next node higher than its
parent and starting again until it gets to the root and goes to the
right side. This time we are checking each level going from the parent
to the left node then right node, then starting again until all of the
right sides of the tree have checked.