Add Priority Queue Class
Add a Priority Queue class that in a Min Binary Heap, parent nodes are
always lower value than child nodes. So, the way we store them is using
an array as heaps have to be as compact as possible; as a result, an
array can stores heaps very close together. Furthermore, a priority
queue is a data structure where each element has a priority. Elements
with higher priorities serve before elements with lower priorities.