Update Hash Function With Second Version
Update Hash Function With Second Version with the following changes:

 - We have add-in Math.min to have an upper limit to the number of times
the loop goes so that if the key is one thousand, it will only loop 100
times in our example. That way, we have some consistency.
 - Most Hash Function uses prime numbers so to reduce the amount of
collations that can happen when two different keys with the same index.
Thus we try to spread out the indexes to avoid those collations. It is
also helpful if the array that you're putting values into has a prime
length. You don't need to know why (Math is complicated!). But here are
some links if you're curious.

Ref:

https://www.quora.com/Does-making-array-size-a-prime-number-help-in-hash-table-implementation-Why