Add Graph Class
Add a graph class for the graph data structures are consist of a set of
nodes together with a set of unordered pairs of these nodes for an
undirected graph or a set of ordered pairs for a directed graph. As long
as a node is connecting to another node, it is considered a graph. There
is also no parent or child nodes. Each node is equal to each other.

Furthermore, an adjacency list is a list of connections using arrays or
hash tables, so the elements in the outer array are for each node in the
graph, and each element has an array with numbers to each node they
connect. We will be using a hash table to store the graph, and it is
going to be an undirected graph.

Which is a graph is with no direction on edge between the nodes, so can
traversal between the nodes both ways.